requirements.txt
# ============================================
#  Python 依赖 - 智能职业发展助手
#  安装命令: pip install -r requirements.txt
#  推荐镜像: pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
# ============================================

# --- Web 框架 ---
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# --- HTTP 请求 ---
requests==2.31.0

# --- 文档处理 ---
PyPDF2>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0

# --- 文本匹配与模糊搜索 ---
thefuzz>=0.22.0
python-Levenshtein>=0.25.0

# --- 数据处理 ---
pandas>=2.0.0

# --- 浏览器自动化 / 爬虫 ---
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# --- HTML 解析 ---
beautifulsoup4>=4.12.0
lxml>=5.0.0

# --- 其他工具 ---
Pillow>=10.0.0
